pyproject.toml
[project]
name = "mechlogic"
version = "0.1.0"
description = "Mechanical logic compiler - converts logic element specs to 3D-printable CAD files"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cadquery>=2.4.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
mechlogic = "mechlogic.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
